feat:add clean code and solid in public folder
